fix: inline columns/DataTable, fix Biome lint errors in appointments UI
- Inline columns and DataTable components into ReceptionistAppointment.tsx
  to avoid Windows/Linux case-sensitivity issues with file renames
- Remove now-redundant columns.tsx and data-table.tsx
- Fix import order in AppointmentsPage, AppointmentDetailsPage,
  ReceptionistAppointment, ReceptionistAppointmentDetails
- Replace array index key with stable composite key in symptoms list
- Convert LocalTime interface to type alias
- Convert CRLF to LF in generated API model files
