+++ frontend/src/api/src/api/model/AppointmentDto.ts
@@ -4,8 +4,8 @@
  * TeleHealth.Api | v1
  * OpenAPI spec version: 1.0.0
  */
-import type { LocalDate } from './LocalDate';
-import type { LocalTime } from './LocalTime';
+import type { LocalDate } from "./LocalDate";
+import type { LocalTime } from "./LocalTime";
 
 export interface AppointmentDto {
   doctorName?: string;
