+++ frontend/src/api/src/api/model/BookAppointmentCommand.ts
@@ -4,7 +4,7 @@
  * TeleHealth.Api | v1
  * OpenAPI spec version: 1.0.0
  */
-import type { Symptom } from './Symptom';
+import type { Symptom } from "./Symptom";
 
 export interface BookAppointmentCommand {
   schedulePublicId: string;
