+++ frontend/src/api/src/api/model/CreateUserCommand.ts
@@ -4,7 +4,7 @@
  * TeleHealth.Api | v1
  * OpenAPI spec version: 1.0.0
  */
-import type { LocalDate } from './LocalDate';
+import type { LocalDate } from "./LocalDate";
 
 export interface CreateUserCommand {
   username: string;
