+++ frontend/src/api/src/api/model/LocalTime.ts
@@ -5,4 +5,5 @@
  * OpenAPI spec version: 1.0.0
  */
 
-export interface LocalTime {}
+// biome-ignore lint/complexity/noBannedTypes: generated file
+export type LocalTime = {};
