+++ frontend/src/api/src/api/model/ReceptionistAppointmentDetailDto.ts
@@ -4,10 +4,10 @@
  * TeleHealth.Api | v1
  * OpenAPI spec version: 1.0.0
  */
-import type { Allergy } from './Allergy';
-import type { LocalDate } from './LocalDate';
-import type { LocalTime } from './LocalTime';
-import type { Symptom } from './Symptom';
+import type { Allergy } from "./Allergy";
+import type { LocalDate } from "./LocalDate";
+import type { LocalTime } from "./LocalTime";
+import type { Symptom } from "./Symptom";
 
 export interface ReceptionistAppointmentDetailDto {
   publicId?: string;
