+++ frontend/src/api/src/api/model/RegisterPatientCommand.ts
@@ -4,7 +4,7 @@
  * TeleHealth.Api | v1
  * OpenAPI spec version: 1.0.0
  */
-import type { LocalDate } from './LocalDate';
+import type { LocalDate } from "./LocalDate";
 
 export interface RegisterPatientCommand {
   username: string;
@@ -14,9 +14,9 @@ export interface RegisterPatientCommand {
   lastName: string;
   icNumber: string;
   /**
-     * @minLength 1
-     * @maxLength 1
-     */
+   * @minLength 1
+   * @maxLength 1
+   */
   gender: string;
   dateOfBirth: LocalDate;
 }
