+++ frontend/src/api/src/api/model/UpdateMedicalRecordCommand.ts
@@ -4,8 +4,8 @@
  * TeleHealth.Api | v1
  * OpenAPI spec version: 1.0.0
  */
-import type { Allergy } from './Allergy';
-import type { EmergencyContact } from './EmergencyContact';
+import type { Allergy } from "./Allergy";
+import type { EmergencyContact } from "./EmergencyContact";
 
 export interface UpdateMedicalRecordCommand {
   bloodGroup: string;
