--- frontend/src/features/appointments/components/data-table.tsx
@@ -1,111 +1,0 @@
-import { useState } from "react";
-import {
-    flexRender,
-    getCoreRowModel,
-    getFilteredRowModel,
-    useReactTable,
-    type ColumnDef,
-} from "@tanstack/react-table";
-import { Search } from "lucide-react";
-import { Input } from "@/components/ui/input";
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow,
-} from "@/components/ui/table";
-
-interface DataTableProps<TData, TValue> {
-    columns: ColumnDef<TData, TValue>[];
-    data: TData[];
-}
-
-export function DataTable<TData, TValue>({ columns, data }: DataTableProps<TData, TValue>) {
-    const [globalFilter, setGlobalFilter] = useState("");
-
-    const table = useReactTable({
-        data,
-        columns,
-        getCoreRowModel: getCoreRowModel(),
-        getFilteredRowModel: getFilteredRowModel(),
-        state: { globalFilter },
-        onGlobalFilterChange: setGlobalFilter,
-    });
-
-    const rows = table.getRowModel().rows;
-
-    return (
-        <div className="space-y-4">
-            {/* Search */}
-            <div className="relative w-72">
-                <Search className="absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-muted-foreground pointer-events-none" />
-                <Input
-                    placeholder="Search appointments…"
-                    value={globalFilter}
-                    onChange={(e) => setGlobalFilter(e.target.value)}
-                    className="pl-9 h-9 text-sm"
-                />
-            </div>
-
-            {/* Table */}
-            <div className="rounded-lg border border-border overflow-hidden">
-                <Table>
-                    <TableHeader>
-                        {table.getHeaderGroups().map((headerGroup) => (
-                            <TableRow
-                                key={headerGroup.id}
-                                className="bg-foreground hover:bg-foreground border-b border-foreground/20"
-                            >
-                                {headerGroup.headers.map((header) => (
-                                    <TableHead
-                                        key={header.id}
-                                        className="text-background/70 font-semibold px-5 py-3.5 text-[11px] tracking-[0.15em] uppercase"
-                                    >
-                                        {header.isPlaceholder
-                                            ? null
-                                            : flexRender(
-                                                  header.column.columnDef.header,
-                                                  header.getContext(),
-                                              )}
-                                    </TableHead>
-                                ))}
-                            </TableRow>
-                        ))}
-                    </TableHeader>
-                    <TableBody>
-                        {rows.length ? (
-                            rows.map((row, i) => (
-                                <TableRow
-                                    key={row.id}
-                                    className="border-b border-border transition-colors duration-100 hover:bg-muted/50"
-                                    style={{ animationDelay: `${i * 30}ms` }}
-                                >
-                                    {row.getVisibleCells().map((cell) => (
-                                        <TableCell key={cell.id} className="px-5 py-3.5 text-sm">
-                                            {flexRender(
-                                                cell.column.columnDef.cell,
-                                                cell.getContext(),
-                                            )}
-                                        </TableCell>
-                                    ))}
-                                </TableRow>
-                            ))
-                        ) : (
-                            <TableRow>
-                                <TableCell
-                                    colSpan={columns.length}
-                                    className="h-32 text-center"
-                                >
-                                    <p className="text-sm text-muted-foreground">No appointments found.</p>
-                                    <p className="text-xs text-muted-foreground/60 mt-1">Try adjusting your search.</p>
-                                </TableCell>
-                            </TableRow>
-                        )}
-                    </TableBody>
-                </Table>
-            </div>
-        </div>
-    );
-}
